Remove TODO about adding an If-Modified-Since header
After a bit more thought, it won't make much difference because we're
always changing the `?since` query param. We're not requesting the same
large resource over & over, we request ever-changing URLs for small
resources, and we do it in 15min intervals.

When we're requesting `readings?latest` or `readings?today` for the
auto-updating UI then this should be revisited.